requirements.txt
pyserial~=3.5
numpy~=2.4.3
pynput~=1.8.1